feat(kded): add getAutoRotate method and rename lockAutoRotate
Summary:
lockAutoRotate(true) would enable the automatic rotation and
lockAutoRotate(false) would disable auto rotation which is the opposite
of what the function name would imply, so rename it to setAutoRotate.

We also need a getter for the value to display the status in the UI.
This getter checks if all applicable outputs have auto rotate enabled
and returns that value.

Test Plan: tested on Plasma Mobile

Reviewers: #plasma, bshah, nicolasfella, romangg

Reviewed By: #plasma, bshah, romangg

Subscribers: romangg, nicolasfella, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28773