daemon: improve consistency of the lid behaviour
We already had some behaviour for the case where the lid closes, we were
not applying it when initialising the daemon though, so a laptop would
start showing the screen to the keyboard until the user takes action.

This change addresses it so that upon initialising we run the same code
we would run if we closed the lid.