Consider global output settings when generating a new config
If we already have global settings for this output, we either ran
the generator for this output so we don't need to find out the best
mode and scale again. Or the user manually changed the config and
selected "For any display arrangement" when saving so we should use
those values here. For example a rotated screen will now stay rotated
in a new config too.