kcm: Use more conventional parent/background for OverlaySheet
Parenting to the KCM itself rather than its parent prevents the view
from disappearing while the OverlaySheet is open, which looks a but odd
and generated a bug report. Let's set the parent to the KCM so that the
rest of the UI beneath the dialog remains visible, as we see in other
software that uses the Kirigami.OverlaySheet component.

BUG: 442563
FIXED-IN: 5.23


(cherry picked from commit 7ca8821b6b704dd99168292994cf0069d42fa37c)