kcm: use less framey and more conventional style for revert dialog
Typically we only put buttons in footers with Kirigami.OverlaySheet
when there is no alternative, typically because the main view is
scrollable. In this case, failing to do so would cause content to
look cut off at the bottom, so we use a defined footer area.

However that is not a design constraint here, so we can use the nore
conventional styling of having the buttons right-aligned at the bottom,
not inside a distinct footer area.