KCM: better handle revert timer and buttons
BUG: 447199
FIXED-IN: 5.23.5

The issue was that the `revertSettings` method was called when the timer timed
out, the settings were reverted, the confirmMsg overlay closed and called a
second time `revertSettings` when the kcm backend has not been reloaded yet.

This makes sure this won't happen, as either the timer of the revert button
will revert the Settings but not simply closing `confirmMsg. This makes the
code simpler by removing book keeping but instead using only the events.