Use more generic name for panel connectors
Kscreen replaces the vendor name for panel connectors replacing it with
a more user facing term "Laptop screen" as the user probably doesn't
know the model of their laptop screen.

The problem is this makes assumptions about the form factor which could
easily be a mobile, tablet and not a laptop.

The name "built-in" conveys as much information to solve the user's task
at hand of identifying which screen is which.