osd: Make the osd service a standalone program
Having the OSD in the daemon is problematic currently as it requires
manual positioning of the window. Because KDED is running arbitrary
other plugins this prevents already porting it to layer shell
in Plasma 5 as only one shell plugin can be used per process currently.
Furthermore in the future positioning will not work because of the Qt6
change to report window positions always in the
corner of the screen. This cannot be opted out globally for the same
reasons.
See also the plasma-workspace commit e87d29c2a31b147d243ce6b9b8d1df7da36f1550