Use XWayland client scaling by default
We discussed potentially turning this on for Plasma 5.26, but wanted to
do more testing first. I as well as other KDE contributors have been
running with it on for the past month with seemingly no drama. Seems
like it's safe enough to turn on by default now.

This commit slightly re-arranges the KCM's UI so that the radio button
corresponding to the default setting remains on top.