loosen the duplication check a tad
we can generally expect the ids to be different unless the manufacturer
hasn't obeyed the spec. with that in mind let's only prime the
duplication detection tech when multiple duplicated outputs are actually
connected. outputs may be temporarily duplicated for reasons other than
duplicated ids. e.g. some devices change output names (DP-1 to DP-2) on
hotplugs and we'd incorrectly handled them as duplicates even when the
old name wasn't connected and bound to go away