kded: Fix crash when connected output reports no modes
When using a KVM switch with EDID emulation that supports two
(or more) screens connected, some switches will report all outputs
as connected, even if only one monitor is plugged into the switch.
In that case, the switch reports one of the outputs as connected
but with no available modes. As a result, the KDED code crashed as
it tried to find the best mode and dereference it without checking
when initializing the output. Since output without any modes cannot
be enabled, we detect that and explicitly disable the output instead.