RFC: sort enabled outputs first
We want the combobox and stackview to show the most relevant outputs
first.

As indexes are passed between here and the top view we sort low in the
stack.

This also saves QML from having to deal with the model object changing.
From a QML POV it'll now be the same object resetting.

> visible: kcm.outputModel.rowCount() > 1

Is ported as this is unsafe. We're calling an invokable as a property so
it won't update.