kcm: make output view more frameless
Right now we have a "frames within frames" effect stemming from the
output panel being a framed view within the main view. This is
unnecessary; we can make it touch the view edged on all sides but the
bottom, and then just add a separator to keep it away from the bottom.