kded: reduce target mobile DPI
192 as a target DPI was too high, generating excessively low scale
factors that resulted in UI elements on screen being too small for the
mobile devices we have autotests for. 136 is better, and results in
larger scales that feel better for these small-screen devices, so you
don't feel like you have to stick your nose right up to the screen to
see anything.