KCM: Fix glitch when disabling north-west output
When toggling south or east positioned outputs, canvas will
automatically shrink or grow as needed. But disabling north-west output
does not trigger repositioning, thus leaving multiscreen canvas in a
weird state with other screens positioned to the right or bottom of an
empty gap.

This fix has a downside that re-enabling north-west output right after
disabling it won't revert the canvas to the state it had before: other
outputs were already repositioned to coordinate {0,0}, and this one
would appear again at the same {0;0} overlapping some of them.