Show applet in tray when relevant
KScreen ships this little applet that's all set up to appear in the
System Tray, and it even includes logic to only appear in the active
area when you have more than one screen connected, and otherwise appear
only in the inactive area. But, the applet is disabled by default, so
none of these things ever happen.

This seems a bit odd; let's change that so that the applet appears in
the inactive area when there's only one screen, and become active with
more than one.