plasmoid: only make active if we're a laptop
Screen layouts rarely change for desktop systems, making this feature
unnecessary there. So for the multi-screen case, ler's only make the
applet active if we're a laptop, where screens tend to be more transient
and there's a higher likelihood of it being useful. Laptops that are
docked 100% of the time are a rarer use case, and not easily detectable,
so for those, users can manually disable this applet if they find that
it annoys them.