kded/output: with duplicate edid hashes, use different global config files
Before this commit, the global config files for two outputs with the same
hash would override each other, making KScreen lose some settings after
a reboot. With this commit, it will detect that and append the connector
name to the file name instead, thus separating the settings for the outputs
properly.

BUG: 452614
BUG: 448599