remove output retention settings
The output retention setting decides whether or not output settings are remembered
per output setup or globally. As the use case for remembering settings such as the
resolution or refresh rate per output setup is incredibly slim to nonexistant, this
commit removes that setting to simplify the UI and code, and makes all settings except
position and priority be global.