Port to upstream QML components
Basically:
- PlasmaExtras.Heading -> Kirigami.Heading
- PlasmaExtras.DescriptiveLabel -> QQC2.Label with manual opacity
- PlasmaComponents3.Label -> QQC2.Label

We can do this because these components only existed in the past to use
special Plasma colors, but no visual styling. And the color stuff is
obsolete because Kirigami colors work correctly in Plasma now.

Everything was tested with a dark Plasma theme and a light color scheme
to ensure no regressions.