Only modify `Coordinate Transformation Matrix`
Allow `libinput Calibration Matrix` to exist as whatever it has set to by default.
This allows for a UDEV rule to be made to set the calibration matrix in cases where
the display and touchscreen do not have the same orientation or pixel sizing.
This now matches the mutter implementation.

BUG: 474110

If my authorship is unsuitable (no real name used), I give permission for you to pick and reauthor at your wish.