kcm/output_model: basic mode matching in mirroring
Provide basic support for adapting display modes for output replication.

Suppose the user wants screen A to replicate screen B. We only change
the configuration of screen A:
1. try to find a candidate mode among all modes with the same aspect
   ratio. Among those, use the smallest mode at least at big as that of
   screen B.
2. Reposition and rescale screen A to fit in the center of screen B.

This always works when screen A has the same mode as the current mode of
screen B. On wayland, this basically works when screen A has modes with
matching aspect ratio.

Since X11 does not support per-screen scaling factors, this works
poorly.