kcm: Move "Identify Screens" button into contextual header actions
And marginally reduce vertical height of the ScreenView. Not by much,
because we still need to allow the user to drag a larger screen
above or below another one. Just a little bit, as the space below
the monitors is now tidier without the extra button in the way.

It looks ever so slightly more elegant this way and reclaims
about half a control's worth of form space.