KCM: Make sure HDR UI elements aren't shown when HDR isn't available
Right now we're assuming that HDR can only be enabled when available,
but that's actually not the case, as HDR can become unavailable later
after having previously enabled it. This can cause the KCM to display
inappropriate UI elements.

Fix that by also adding an "Is HDR available?" check to the visibility
conditions that currently only check whether HDR is currently enabled.

BUG: 482748
FIXED-IN: 6.0.2