osd: Dont quit before config is applied
Fixes osd doing nothing on X11. It worked by accident on Wayland
because we run one more event loop iteration before quitting.
This makes it so we only quit when the SetConfigOperation is finished
and no other SetConfigOperation is in flight and the user requested
not another round of actions.