kcm: set a11y name and role for orientation buttons
Screen orientation in the kcm is configured through a set of icon-only
toolbuttons in a buttongroup with a checked property. This has two
negative a11y implications: first, there is no label to read out for
these buttons. Second, their type is read out as checkbox, but as screen
orientations are mutually exclusive (per device) this is semantically
ill-fitting.

This change sets the Accessible.name to the tool-tip for the button,
which is reasonably clear and short, and sets Accessible.role to
the semantically more adequate RadioButton.