kcm: make representation in sliders percentages
Screen scaling can be configured through either a slider or a spinbox.
The internal representation of scaling is as a direct multiplier, but
to the user this is displayed as a percentage. Screen readers, however,
do not pick up on this translation in the case of the slider. The values
read out when initially focusing the contol and when moving the slider
are the internal values, inappropriately rounded to one decimal digit;
e.g., a value presented in the UI as 125% is read out as 1.2.

This change modifies the sliders to use percent values as their internal
representation, translating them as needed from/to the values expected
in the kcm. The slider's accssible description further explains the
value.