kcm: Add mobile/touch interface
The KCM UX currently is not very usable on mobile. This MR experiments
with adding a second interface implemented with FormCard for mobile
purposes (only shown when QT_QUICK_CONTROLS_MOBILE=1)

Big TODO, currently I made a clone of the output panel which has a ton
of copy pasted code for every control. I need to figure out how to share code between the
two UIs...