kcm: Add mnemonics
Kirigami FormLayout buddy defaults to the item it is attached to.
However, in KScreen we often have a RowLayout to place the contextual
help button. Therefore, it did not add mnemonics automatically.

Explicitly set the buddy to the relevant control.

While at it, give explicitly give "HDR" the Alt+H shortcut.