osd: workaround wrong geometry on X11
#### Description

A bug is raised on X11 - on the first display, the OSD is displayed with the wrong geometry and is not exactly in the center. On the second showing to `showActionSelector()`, if the on-screen menu is visible, the geometry is adjusted to center.

<details><summary>Screenshots</summary>
![first](/uploads/8fea80b4169a4965e742365f3dbbbb1c/изображение.png)
![second](/uploads/6feda602fc84e6fcd32d795504fcb29c/изображение.png)
</details>

There is no such problem on Wayland.

#### Analysis and solution

I found out that it's related to the type of window `NET::OnScreenDisplay`. For example, if replace it with `NET::Notification`, then this problem no longer arises.

Maybe, like it says here, https://doc.qt.io/qt-6/application-windows.html#window-geometry, this is some kind of X11 pitfall.

At least, that's how I understand it, in this case with OSD, the non-standard value `_KDE_NET_WM_WINDOW_TYPE_ON_SCREEN_DISPLAY` is set. And probably because of this, there is some kind of problem when setting the geometry before the window is displayed (visible).

I also noticed that if we set the geometry (`setGeometry()`) after the OSD is visible `setVisible(true)`, then there is no problem.

Therefore, such a workaround is proposed here, but if a more elegant solution is possible, for example in KWindowSystem, then I don't mind.


(cherry picked from commit b3915130d3d24200e08ac8aae7af62dd45470afd)

a2d81cfd osd: workaround wrong geometry on X11

Co-authored-by: Sergey Katunin <sulmpx60@yandex.ru>