osd: moving selected action on repeated showing
At the moment, when the display switching menu is called again (for example, by pressing `Win+P`), there is no movement in the list of actions. 

However, it would be convenient, and this is how it works on Windows OS, for example.

This change adds a move to the right on repeated showing of OSD, by invoking new `moveRight` QML function.