Fix UB in sorting of KScreen modes
Previous sorting comparator used a comparison that is not a strict
ordering. It is UB to use it in std::sort.

There is no way to say which screen is "the biggest", as one screen can
be wider, while another is taller. Hence, instead, the widest screen is
selected as the biggest one (tallest if there is a tie).