kcm: Add checkbox to disable DDC/CI for a given output
Requires updates to plasma-wayland-protocols and libkscreen to
build, as well as updates to kwin and powerdevil to work as intended.

Some monitors have buggy DDC/CI implementations, or the user may
want to disable it for other reasons. Allow them to do so and use
KWin's software brightness controls instead.

In the future, this toggle may also affect contrast settings,
input source selection, or any other useful DDC/CI features.