kcm: Allow opening a specific screen on start-up
The KCM sensibly selects the first available, enabled screen when
opened. However, there are cases - such as right-clicking the desktop,
or trying to configure the built-in screen of a device - where we want
to save a few clicks and select a specific screen by default.

This is especially useful in multi-monitor scenarios where the screen is
badly named, or has the same name as many other screens.

To open a specific screen, you pass the *priority index* (which is used
in libplasma and exposed in Qt applications in that order), not as seen
the screens are ordered in the KCM.