kcm: Reload rotation for disabled outputs on X11
Similar to 5964eabc69c369b9c727d21199ef196a4c7d00e2 in the osd.
On X11 the roation for disabled outputs is lost. When a rotated
output is disabled and enabled again, the UI would still show
it as rotated but the geometry and the result was non-rotated.
In this case make sure to set the rotation when enabling it again.
Also load the the last set rotation value for disabled outputs
as a convenience for the user.