kcm: fix handling of replication source indices
We can't simultaneously filter outputs from the replication source model, and also use the
index of an output in that model with the normal outputs list.
To fix that, create a filtered list of outputs as possible mirror sources and use that to create
the replication source model *and* to address the output by its index.