hdrcalibrator: add a summary page with a setting for Windows HDR apps
Windows games are a bit special with their HDR support, as most of them do not
follow system configuration but have their own calibration settings.
Because these calibration settings assume the system behaves like MS Windows
and just copies the game contents to the screen unchanged, they don't work
particularly well in Plasma.

To work around that, we need to assume that these applications target some
predefined settings. The added option allows the user to choose which screen
or rather which calibration Windows applications on the system can be assumed
to be calibrated towards.